refactored core part of graphEngine to share functionalities between OTG and IMG
