+++ src/core/engine/OTG.ts
@@ -1,17 +1,9 @@
-import { connectionFinder } from "../connections/ConnectionFinder";
-import { ConnectionDetails, EngineConnections } from "../connections/EngineConnections";
-import { extractor, NodePorts } from "../connections/Extractor";
+import { EngineConnections } from "../connections/EngineConnections";
 import { ConfigNodeDict } from "../nodes/ConfigNode";
 import { EngineNodeDict } from "../nodes/EngineNode";
-import { LogicNode, LogicNodeDict } from "../nodes/LogicNode";
 import { nodeConverter } from "./Converter";
+import { executeNode, GraphExe } from "./Core";
 import { validator } from "./Validator";
-
-export interface GraphExe {
-    nodes: LogicNodeDict,
-    connections: EngineConnections,
-    entry: string
-}
 
 /**
  *  One time graph will be triggered once and destroyed after successful execution 
@@ -44,44 +36,6 @@ export const executeGraph = (
         if (node.onDestroy)
             node.onDestroy(...node.inputs, ...node.outputs)
     })
-}
-
-/**
- *  Resolves all dependencies (input ports) for this node based on the connections
- * @param node 
- * @param otg 
- */
-const resolveDependency = (node: LogicNode, otg: GraphExe) => {
-
-    const ports: NodePorts = extractor(node);
-
-    ports.inputs.forEach((con) => {
-
-        const dependencies: ConnectionDetails[] = connectionFinder(con, otg.connections);
-
-        dependencies.forEach(dep => {
-
-            //execute the dependencyNode and set it's ioPorts value
-            executeNode(otg.nodes[dep.nodeId], true, otg);
-
-            //assign the computed value to this input now
-            node.inputs[con.index].value = otg.nodes[dep.nodeId].outputs[dep.index].value;
-
-        })
-
-    })
-
-}
-
-/**
- * Executes the logic of a node
- * @param node The to be executed node
- * @param isDependency Handles of output nodes should be fired form this node 
- */
-const executeNode = (node: LogicNode, isDependency: boolean, otg: GraphExe,) => {
-    resolveDependency(node, otg);
-    if (isDependency && !node.autoUpdate) return;
-    node.exe(...node.inputs, ...node.outputs);
 }
 
 const createOTG = (
